backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg[binary]==3.1.18  # Modern psycopg3 with binary package (no build required)

# Redis
redis==5.0.1
hiredis==2.3.2
# Note: redis-om and redistimeseries have compatibility issues with redis 5.x
# We'll use redis client directly for TimeSeries commands (TS.CREATE, TS.ADD, etc.)
# redis-om>=0.3.0  # Redis Object Mapper - incompatible with redis 5.x
# redistimeseries>=1.4.5  # Redis TimeSeries client - incompatible with redis 5.x

# Neo4j
neo4j==5.15.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
# FastAPI includes CORS middleware by default (fastapi.middleware.cors)

# WebSockets
websockets==12.0

# Background tasks
celery==5.3.6
rq==1.15.1  # Python RQ (Redis Queue) - simpler alternative to Celery

# Utils
pytz==2024.1
zstandard==0.22.0  # Decompress .zst files from Databento
psutil==5.9.5

# AI Assistant
anthropic==0.39.0  # Claude API
google-generativeai==0.8.3  # Gemini
mem0ai==0.1.24  # Memory for AI agents
langgraph==0.2.45  # Orchestration
langchain==0.3.7  # For SQL agent utilities
vanna[chromadb,anthropic]==0.8.7  # NL to SQL with RAG
chromadb==0.5.18  # Vector DB for Vanna
